Support for Decoding AIS Messages Which Span Multiple NMEA Sentences
If AisDecode is passed an array instead of a string, it will assume that
the sentences in the array make up the multiple parts of a NMEA encoded
AIS message.

The sentences in the array must be in order, must be the complete set
required to reassemble the AIS message and must correspond to no more
than one AIS message. It is the responsibility of the caller to ensure
these constraints are met.